feat: chef delete

+++ src/app/models/File.js
@@ -22,20 +22,22 @@ module.exports = {
 
     return results.rows[0]
   },
-  async delete(id) {
+  async delete(id, isRecipe) {
     try {
       const result = await db.query(`SELECT * FROM files WHERE id = $1`, [id]);
       const file = result.rows[0];
 
       fs.unlinkSync(file.path);
 
-      await db.query(`
-      DELETE FROM recipe_files WHERE file_id = $1
-    `, [id])
+      if (isRecipe) {
+        await db.query(`
+          DELETE FROM recipe_files WHERE file_id = $1
+        `, [id])
+      }
 
       return db.query(`
-      DELETE FROM files WHERE id = $1
-    `, [id])
+        DELETE FROM files WHERE id = $1
+      `, [id])
     } catch (err) {
       console.error(err);
     }
